move:#18 연차 관련 상수 파일 변경
